Rename hamilt_ofdft to module_ofdft
